+++ zdzc-common/src/test/resources/template/controller-restful.ftl
@@ -5,7 +5,7 @@ import ${basePackage}.model.${modelNameUpperCamel};
 import ${basePackage}.service.I${modelNameUpperCamel}Service;
 import org.springframework.web.bind.annotation.*;
 import com.zdzc.common.PageList;
-
+import org.springframework.util.StringUtils;
 import javax.annotation.Resource;
 
 
@@ -27,7 +27,10 @@ public class ${modelNameUpperCamel}Controller {
 
     @DeleteMapping
     public int delete(@RequestParam("id") String id){
-       return ${modelNameLowerCamel}Service.deleteById(id);
+        if(StringUtils.isEmpty(id)){
+            throw new BaseException(ExceptionEnum.SYSTEM_PARAMSID_NULL);
+        }
+        return ${modelNameLowerCamel}Service.deleteById(id);
     }
 
     @PutMapping
@@ -37,6 +40,9 @@ public class ${modelNameUpperCamel}Controller {
 
     @GetMapping("/findById")
     public ${modelNameUpperCamel} detail(@RequestParam("id") String id){
+        if(StringUtils.isEmpty(id)){
+            throw new BaseException(ExceptionEnum.SYSTEM_PARAMSID_NULL);
+        }
         return ${modelNameLowerCamel}Service.findById(id);
     }
